refactor:simplified instantiation of OID4VCIssuerEndpoint
